💄 試合情報修正と駐車場リストのカード化
