+++ src/components/ParkingList.tsx
@@ -1,45 +1,15 @@
 import type { VFC } from "react";
-import {
-  Stack,
-  StackDivider,
-  useColorModeValue,
-  Heading,
-  Flex,
-  Spacer,
-  Box,
-} from "@chakra-ui/react";
-import { ChevronRightIcon } from "@chakra-ui/icons";
+import { Stack } from "@chakra-ui/react";
 import data from "@/data/data.json";
-import { useRouter } from "next/router";
+import ParkingCard from "@/components/ParkingCard";
 
 const ParkingList: VFC = () => {
-  const router = useRouter();
   const parkings = data.parkings;
 
   return (
-    <Stack
-      spacing={5}
-      divider={<StackDivider borderColor={useColorModeValue("gray.100", "gray.700")} />}
-    >
+    <Stack spacing={2}>
       {parkings.map((parking) => (
-        <Box
-          key={parking.key}
-          onClick={() => {
-            router.push({
-              pathname: "/parking",
-              query: { parking: parking.key },
-            });
-          }}
-          cursor="pointer"
-        >
-          <Flex>
-            <Heading as="h5" size="sm" pl="3">
-              {parking.name}
-            </Heading>
-            <Spacer />
-            <ChevronRightIcon mr="3" />
-          </Flex>
-        </Box>
+        <ParkingCard key={parking.key} parking={parking} />
       ))}
     </Stack>
   );
